Foram adicionados os botões Excluir e Editar no caso de uso 003.

+++ src/main/java/br/ufrn/cerescaico/bsi/sigest/dao/EmpresaJpaController.java
@@ -79,7 +79,7 @@ public class EmpresaJpaController implements Serializable {
         }
     }
 
-    public void destroy(BigDecimal id) throws NonexistentEntityException {
+    public void destroy(Integer id) throws NonexistentEntityException {
         EntityManager em = null;
         try {
             em = getEntityManager();
